refactor(gui-app): extract main.rs view ops to runtime_camera_pane.rs (2b shard 1) + clippy cleanup
First decomposition shard of the gui-app monolith (source-size governance /
decision 021). Behavior unchanged.

- Extract the camera/pane/dock-resize view-ops cluster from main.rs's `impl
  Runtime` into a new sibling module runtime_camera_pane.rs (fit/pan/zoom camera,
  workspace-pane focus/split/close/zoom/preset, divider-drag resize, layout &
  hit-test helpers, dock resize drag). `use super::*` + `impl Runtime` mirrors the
  existing runtime_terminal_context.rs pattern; 26 methods promoted to pub(super)
  to preserve their crate-root visibility. main.rs 3942 -> 3526.
- specs/source_module_size_manifest.json: ratchet main.rs recorded_lines to 3526
  and its ceiling down to 4055 with the shrink (the decomposition-only path that
  lowers a frozen ceiling).
- clippy: auto-fix the machine-applicable lints across gui-app (boolean
  simplifications, redundant returns/closures, map_or, into_iter, ref-deref,
  Range::contains, if-collapse) so main.rs and siblings are clippy-clean. The 5
  remaining "too many arguments" warnings are pre-existing, not auto-fixable, and
  outside main.rs.

Full GUI suite green (214 gui-app tests); build + governance gates green.

+++ crates/gui-app/src/terminal_session.rs
@@ -466,7 +466,7 @@ pub(super) fn refresh_terminal_session_context(
         model_revision: context.source_revision.clone(),
         created_unix_ms: read_session_created_unix_ms(&session.session_path)
             .unwrap_or_else(|| unix_time_ms().unwrap_or(0)),
-        process_group_id: Some(session.process_group_id as i32),
+        process_group_id: Some(session.process_group_id),
     };
     write_terminal_context_files(&terminal_context, context)
 }
@@ -522,19 +522,19 @@ pub(super) fn mark_terminal_session_lifecycle(
         &session.context_path,
         lifecycle,
         process_exit_code,
-        Some(session.process_group_id as i32),
+        Some(session.process_group_id),
     )?;
     update_terminal_lifecycle_file(
         &session.latest_context_path,
         lifecycle,
         process_exit_code,
-        Some(session.process_group_id as i32),
+        Some(session.process_group_id),
     )?;
     update_terminal_lifecycle_file(
         &session.session_path,
         lifecycle,
         process_exit_code,
-        Some(session.process_group_id as i32),
+        Some(session.process_group_id),
     )
 }
